added results view init
