+++ src/components/CompareResults/ResultsView.tsx
@@ -15,12 +15,12 @@ import useHandleChangeSearch from '../../hooks/useHandleChangeSearch';
 import { updateFramework } from '../../reducers/FrameworkSlice';
 import { SearchContainerStyles } from '../../styles';
 import { background } from '../../styles';
-import { fetchRecentRevisions } from '../../thunks/searchThunk';
-import { Repository, View, InputType } from '../../types/state';
+import { Repository, View } from '../../types/state';
 import { Framework } from '../../types/types';
 import CompareWithBase from '../Search/CompareWithBase';
 import PerfCompareHeader from '../Shared/PerfCompareHeader';
 import ResultsMain from './ResultsMain';
+import ResultsViewInit from './ResultsViewInit';
 
 interface ResultsViewProps {
   protocolTheme: Theme;
@@ -55,10 +55,7 @@ function ResultsView(props: ResultsViewProps) {
     baseRepos: selectedBaseRepositories as Repository['name'][],
     newRepos: selectedNewRepositories as Repository['name'][],
   };
-  const repositoryBase = useAppSelector(
-    (state) => state.search.base.repository,
-  );
-  const repositoryNew = useAppSelector((state) => state.search.new.repository);
+
   const { dispatchFetchCompareResults, dispatchFakeCompareResults } =
     useFetchCompareResults();
   const { searchByRevisionOrEmail } = useHandleChangeSearch();
@@ -133,29 +130,6 @@ function ResultsView(props: ResultsViewProps) {
     }
   }, [framework]);
 
-  /* editing the revisions requires fetching the
-   * recent revisions in results view
-   */
-  useEffect(() => {
-    const repository = repositoryBase;
-    void dispatch(
-      fetchRecentRevisions({
-        repository,
-        searchType: 'base' as InputType,
-      }),
-    );
-  }, [repositoryBase]);
-
-  useEffect(() => {
-    const repository = repositoryNew;
-    void dispatch(
-      fetchRecentRevisions({
-        repository,
-        searchType: 'new' as InputType,
-      }),
-    );
-  }, [repositoryNew]);
-
   return (
     <div
       className={styles.container}
@@ -173,7 +147,7 @@ function ResultsView(props: ResultsViewProps) {
             <p>Home</p>
           </Stack>
         </Link>
-
+        <ResultsViewInit />
         <CompareWithBase
           mode={themeMode}
           isEditable={true}
